Fix: Move classes into a different namespace
