+++ src/Definition.php
@@ -11,9 +11,8 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Definition;
+namespace Ergebnis\FactoryBot;
 
-use Ergebnis\FactoryBot\FixtureFactory;
 use Faker\Generator;
 
 interface Definition
