+++ test/Fixture/Definitions/DoesNotImplementDefinition/RepositoryDefinition.php
@@ -11,20 +11,14 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ImplementsDefinitionButThrowsExceptionDuringConstruction;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\DoesNotImplementDefinition;
 
-use Ergebnis\FactoryBot\Definition\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 use Faker\Generator;
 
-final class RepositoryDefinition implements Definition
+final class RepositoryDefinition
 {
-    public function __construct()
-    {
-        throw new \RuntimeException();
-    }
-
     public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
     {
         $fixtureFactory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
