+++ test/Fixture/Definitions/ImplementsDefinitionButHasPrivateConstructor/RepositoryDefinition.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ImplementsDefinitionButHasPrivateConstructor;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\ImplementsDefinitionButHasPrivateConstructor;
 
-use Ergebnis\FactoryBot\Definition\Definition;
+use Ergebnis\FactoryBot\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 use Faker\Generator;
