+++ test/Fixture/Definitions/ImplementsDefinitionButIsAbstract/RepositoryDefinition.php
@@ -11,16 +11,16 @@ declare(strict_types=1);
  * @see https://github.com/ergebnis/factory-bot
  */
 
-namespace Ergebnis\FactoryBot\Test\Fixture\Definition\Definitions\ImplementsDefinition;
+namespace Ergebnis\FactoryBot\Test\Fixture\Definitions\ImplementsDefinitionButIsAbstract;
 
-use Ergebnis\FactoryBot\Definition\Definition;
+use Ergebnis\FactoryBot\Definition;
 use Ergebnis\FactoryBot\FixtureFactory;
 use Ergebnis\FactoryBot\Test\Fixture;
 use Faker\Generator;
 
-final class RepositoryDefinition implements Definition
+abstract class RepositoryDefinition implements Definition
 {
-    public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
+    final public function accept(FixtureFactory $fixtureFactory, Generator $faker): void
     {
         $fixtureFactory->defineEntity(Fixture\FixtureFactory\Entity\Repository::class);
     }
